Fix version Add Sms-Uslugi service
