+++ composer.json
@@ -1,8 +1,7 @@
 {
     "name": "omnilight/yii2-sms",
     "description": "Sms sending extension",
-    "keywords": ["yii", "sms"],
-    "minimum-stability": "dev",
+    "keywords": ["yii2", "sms"],
     "authors": [
         {
             "name": "Pavel Agalecky",
@@ -12,10 +11,10 @@
     "type": "yii2-extension",
     "require": {
         "php": ">=5.4.0",
-        "yiisoft/yii2": "dev-master",
-        "guzzlehttp/guzzle": "~5.0"
+        "yiisoft/yii2": "^2.0",
+        "guzzlehttp/guzzle": ">=5.0"
     },
     "autoload": {
-        "psr-4": { "omnilight\\sms\\": "" }
+        "psr-4": { "omnilight\\sms\\": "src/" }
     }
 }
